anomaly-detector/requirements.txt
prometheus-api-client==0.5.3
scikit-learn==1.3.2
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0
statsmodels==0.14.0
